requirements.txt
Django>=5.0.1,<5.1
djangorestframework>=3.14.0,<3.15
markdown>=3.5.2,<3.6
django-filter>=23.5,<23.6
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.27.0,<0.28
Pillow>=10.2.0,<10.3
uwsgi>=2.0.23,<2.1
strawberry-graphql-django>=0.32.0,<0.33
django-choices-field>=2.2.2,<2.3